iproto: fix multiple close of same socket
All iproto threads listening same socket, and if user change listen
address, this socket closed in each iproto thread. This patch fix
this error, now socket closes only in main thread, and in other threads
we are only stop listening, without socket closing. Also this patch fix
error, related to the fact that the tarantool did not delete the unix
socket path, when it's finishing work.