alter: fix rollback of replace trigger for _user
While on_replace_dd_user() trigger is called for update/replace
operation, we firstly apply operation and then setup rollback trigger in
case transaction will be rolled back/aborted.

on_replace_dd_user() takes two tuples: old and new ones. As a rule we
assume that if (old != NULL && new != NULL) then the operation is
replace. However, for _user space that's not the case. In fact, even if
old == NULL we still make lookup by new->id in user_cache and if user
already exists with this is, then we account operation as replace.
That's handled this way due to the fact that some entries (to be more
precise 'admin' and 'guest') are inserted into user cache before we
start recovery from snapshot.

Finally, before this patch, while we are arming on_rollback trigger for
replace operation, we pass old tuple as an argument. However, as it was
mentioned before, old tuple can be absent. That said, it may lead to
crash. Let's fix it in the following way: now we are passing not tuple,
but old user definition to the rollback trigger. Moreover, we have to
copy it since replace in user cache results in memory release of old
definition. And in order to released duplicated old user definition, we
arm on_commit trigger. That's it.

Closes #5998