sql: show varbinary in type mismatch error
Currently, the word "varbinary" is printed instead of the value in the
description of the type mismatch error. This is due to the problem of
outputting binary values. However, this sometimes leads to confusion
because in most other cases the value is printed. After this patch,
instead of the word "varbinary", the actual value will be output in
hexadecimal format in the description of the type mismatch error.

Follow up #4356