sql: make type mismatch errors more informative
Prior to this patch, in some cases the type mismatch error description
showed the value, and in some cases the type of the value. After this
patch, both the type and value will be shown. Also, inconsistent type
error description also become more informative. Previously it contained
only type of value, now it contains value and its type.

Close #6176